# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apartments API",
        "description": "",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/apartment": {
            "get": {
                "tags": [
                    "Apartments"
                ],
                "description": "Get List of apartments",
                "operationId": "6a5296ebbabc430c8c3eb850e4113d90",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting parameters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name:asc,price:desc,created_at:desc"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter Apartment Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Filter Apartment price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "Filter Currency of price",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Filter by size",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        }
                    },
                    {
                        "name": "balcony_size",
                        "in": "query",
                        "description": "Filter by balcony_size",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filter by Description",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by ID category_id ",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Apartments"
                ],
                "description": "Get List of apartments",
                "operationId": "a1ef49060bfcbbc2b7e09c8199572d42",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Apartment Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Apartment price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "Currency of price",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "properties",
                        "in": "query",
                        "description": "Apartment Properties",
                        "required": false,
                        "schema": {
                            "type": "json"
                        },
                        "example": "{'size':'numeric','balcony_size':'numeric','location':'string'}"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Apartment Description",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID of category to which  apartment belongs",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Unprocessable Entity - validation failed"
                    }
                }
            }
        },
        "/apartment/{id}": {
            "put": {
                "tags": [
                    "Apartments"
                ],
                "description": "Get List of apartments",
                "operationId": "45fb1f343030ddb802c5cf898cc6ccf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Apartment Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Apartment price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "Currency of price",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "properties",
                        "in": "query",
                        "description": "Apartment Properties",
                        "required": false,
                        "schema": {
                            "type": "json"
                        },
                        "example": "{'size':'numeric','balcony_size':'numeric','location':'string'}"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Apartment Description",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID of category to which  apartment belongs ",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Unprocessable Entity - validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Apartments"
                ],
                "description": "Delete apartment",
                "operationId": "65ca503de459acaad5d9ffaf6e7d9c20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}